back/requirements.txt
anyascii==0.3.1
appdirs==1.4.4
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.1
black==22.12.0
certifi==2022.12.7
chardet==5.1.0
Click>=8.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.6.0
djangorestframework==3.14.0
draftjs-exporter>=2.1.5,<3.0
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
openpyxl==3.0.10
Pillow==9.4.0
pytz==2022.7.1
qrcodegen==1.8.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
telepath==0.3
toml==0.10.2
Unidecode==1.3.6
urllib3==1.26.14
wagtail==4.1.1
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.7
xlwt==1.3.0
